requirements.txt
google-generativeai==0.7.2
pydantic==2.8.2
pydantic_settings==2.4.0
python-telegram-bot==21.4
